+++ CHANGELOG.md
@@ -2,6 +2,8 @@ DevOps Tool: Core
 ==============================================
 
 # 0.1.1 (Unreleased)
+- Added DatabaseAdapterManager
+- Merged DatabaseMetadataProviderInterface into DatabaseAdapterInterface
 - Added exception handling around entire app
 - Added FilesystemAdapterManager
 - Added filesystem:ls, filesystem:copy, and filesystem:sync commands